requirements.txt
whitenoise
django-widget-tweaks